Patched fact and manifests to preserve partitions
The ceph::osd class deploys the storage in two steps: prepare partitions
and then activate their. This class uses devices provided by the
osd_devices_list fact. This fact looks on the GUID type of partitions
and lists them. After that the manifest sequentially executes two
command like that:
	ceph-deploy osd prepare <devices>
and then:
	ceph-deploy osd activate <devices>

When we want to preserv partitions for OSDs the prepare command should
be skipped for them. This patch contains a modified fact, the new fact
returns two lists of partitions in the result string that separated by
';'. The first list contains partitions that should be prepared by the
ceph-deploy tool. And the second list contains partitions that should be
activated by this tool. How does this fact place partitions in these
lists? If a mounted partition does not contain the fsid file, then
this partition should be prepared.

There is update.sh that simplifies copying of file into /etc/puppet/.
The original puppet manifests located on the master node and distributed
by rsync.