Get IP addresses to replace from node's deploy info
Originally we take IPs of CICs in 6.0 env from deploy info for primary
controller. However, after it is deployed, it's deploy info contains replaced IP
addresses and cannot serve as a template.
So provide filename as an argument to replace ips functions and pass role-based
filename to them to get 6.0 IP addresses from.