Patch Puppet modules consistently
All patches to Puppet modules are kept in octane/patches/puppet directory. For
every module to be patched, there is a subdir of the same name as the module.
This subdir contains 'patch' file. Script octane/patches/puppet/update.sh
applies all patches to /etc/puppet/2014.2.2-6.1/modules dir.

Patches are created with 'git diff' on the fuel-library repository.