Set port to UP after adding it to bridge while upgrade controllers
To isolate new controllers from original ones in the same network, we disable
actions that add ports to bridges br-ex and br-mgmt. However, in this case
Puppet doesn't set the port up, and when we upgrade controller, and reconnect it
to physical network, connection isn't established.

Add 'ip link set up' command for the port device after adding it to bridge
during replacement of controllers.