Merge deployment info
To update deployment info for the env without losing data, we must upload
modified facts for all nodes every time we change anything in a single node.

Add function to get default and current facts for the upgrade seed env and
combine them in a single list. Do this every time we run
ControllerUpgrade.postdeploy on each controller node.

Change-Id: I26b9dbd196eabd4a99a95facc9b7aa459c527a20