Remove unchanged network groups from uploaded data
Setting of network groups configuration for the cluster updates configuration of
all network groups passed to API. Updated network groups are deleted and
recreated. This leads to delete of VIPs, and, in case of cloned env, new VIPs
will differ from the oritinal ones.

Remove unmodified network groups from data to be uploaded to API to avoid
deleting those network groups and VIPs assocciated with them.

Change-Id: I513e563f02606b26fe341151dd9ef8bb81d08e34