Find Nova service tenant ID file by ID of source env
The bug causes update_neutron_config function in upgrade_controlplane module
look for ID of service tenant in a file with an id of seed env in its name.
However, the file is created with id of original env.
This bug won't affect scenario when sync-images command is run, because it will
create the file with id of seed env.

Change-Id: I5ae4aa39099084763a6f1d702e7d4131d69710ed