Set timeout for restarting containers via systemd
In systemd, default timeout for starting/stopping operation
is 90s. This is insufficient for operations on Fuel contianers
as they run Puppet manifests and/or database sync before the
service actually started.

Override default value for the timeout via configuration file
in /etc/systemd/ directory. Delete the file after the container
is restarted.

Change-Id: I0fd5879c4817092400ced1b48b6ac9bd1613d07e