Remove docker related code from AstuteArchivator
Since the removing docker containers from the host re-initialization of
services should be done by applying puppet manifests. The code that
responsible for doing this with containers were removed from
AstuteArchivator and now this archivator responsible only for restoring
the astute.yaml file, this re-initialization will be added as a separate
archivator.

Implements: blueprint upgrade-fuel-admin-node
Change-Id: Ied3b5132e9dce1c37b259f74d76c93f56dbcdf67