Remove fixes for partitioning schema for versions before 7.0
Partitions schema created by fuel-agent for image-based provisioning differs
from the schema used for cobbler provisioning. Cobbler provisioning was
supported in versions prior to 7.0.

If version of the environment is lower than 7.0 and provisioning method is not
image-based, create config-drive partition at the node being upgraded. Remove
scripts that modify partitions metadata on nodes and change the way the disk
partitions sizes are calculated.

Change-Id: I27a5477186efe62e3e3dc6f96b16889da1a49250
Closes-bug: 1574624