Allow Octane to use non-root user and sudo.
Adding 4 options to Octane:
    --ssh-user
    --ssh-password
    --ssh-password-inline
    --sudo

to allow Octane to SSH as a non-root user and prepend all commands
with `sudo`.  It will also spawn `sftp-server` using sudo so that
the SFTP client in Paramiko will work as-is.

And one new command: `octane check-sudo`

This command will log into the specified nodes (or all nodes if none
are passed) and attempt to SSH and use `sudo`, as well as test that
SFTP works using `sudo`.

Change-Id: I0560587c8f27447c7683e5a8fd2b643730d83bce