Remove fix for custom field in release metadata
Remove the script that patches custom field in the release metadata from
the code tree.

This script was requried for upgrade of customized version of 5.1.1
release of Fuel. Since that version is no longer supported, the fix
is also not used anymore.

We still need to prepare for later patching of a node's metadata by
moving other scripts and patches to the location where they are expected
to be.

Change-Id: Ifb2590f96449716db09219240ceda6738951f95b
Related-bug: 1574624