Improve partitions and disk metadata handling
Partitions schema and disk metadata at nodes must be updated only
if they are scheduled for reprovisioning, and the provisioning
method has changed to incompatible (i.e. from 'cobbler' to 'image')
due to the shift of release version.

Move logic that manages partitions from upgrade handlers to function
that actually initiates the re-provisioning of node. This ensures
that for every node this is applied only once and only if necessary.

Change-Id: I28718564543e3ccd92d985bd79c76e509e8eae1f
Closes-bug: 1585938
Closes-bug: 1575054