Remove containers specifics from DB archivators
Since nailgun and keystone services are not isolated in containers their
processes should be stopped before to manipulate with databases and
data.

DB archivators are collected in DatabasesArchivator to apply the
'postgresql' task before to restore DB states. This task is needed to
configure databases, users and privileges according new parameters in
astute.yaml.

Some new functions were introduced to perform respective operations on
the host level, such as:
  - archivate.archivate_cmd_output
  - sql.run_psql

Change-Id: I67145808e8c453e371fe20575c5cc6a729db139a