Revert admin_token_auth from keystone pipelines
The presence of the admin_token_auth middleware is needed only to apply
the keystone puppet task during the restore phase. This middleware
allows to authorize with admin_token and configure things in Keystone.
After the successful re-initialization of Keystone this middleware
should be removed from pipelines alongside with restarting the keystone
service because of security reasons.

Change-Id: Id98d8f25270538ab850af936eff749f0277e0c58