Restart mcollective on slave nodes after restore
Also, on the backup step a status of `mco ping` is serialized in an
upgrade tarball and on the restore step is compared with the actual
status. All nodes that are not respond are logged.

In additional, the status of `mco ping` is archived on the backup step
and is compared on the restore step with the actual ones.

Change-Id: Ibba81102214998d83614a42cdb21c21bebd8284a
Related-Bug: #1561092